refactor(deployment): fix broken imports
