+++ lib/manager/components/ProjectSettings.js
@@ -6,7 +6,7 @@ import {LinkContainer} from 'react-router-bootstrap'
 
 import {deleteProject, updateProject} from '../actions/projects'
 import {getComponentMessages, isModuleEnabled} from '../../common/util/config'
-import DeploymentSettings from './DeploymentSettings'
+import DeploymentSettings from './deployment/DeploymentSettings'
 import ProjectSettingsForm from './ProjectSettingsForm'
 
 import type {Project} from '../../types'
