+++ lib/manager/components/version/FeedVersionNavigator.js
@@ -16,7 +16,7 @@ import {isValidZipFile} from '../../../common/util/util'
 import * as snapshotActions from '../../../editor/actions/snapshots'
 import * as gtfsPlusActions from '../../../gtfsplus/actions/gtfsplus'
 import * as deploymentActions from '../../../manager/actions/deployments'
-import DeploymentPreviewButton from '../DeploymentPreviewButton'
+import DeploymentPreviewButton from '../deployment/DeploymentPreviewButton'
 import FeedVersionViewer from './FeedVersionViewer'
 import VersionSelectorDropdown from './VersionSelectorDropdown'
